measure input timestamps in ticks

+++ wootosu.Shared/Models/RecordFrame.cs
@@ -6,7 +6,7 @@ public class RecordFrame
 {
   public RecordFrame() { }
 
-  public RecordFrame(long scanIndex, int timestamp, VirtualKey keyCode, float analogValue)
+  public RecordFrame(long scanIndex, float timestamp, VirtualKey keyCode, float analogValue)
   {
     ScanIndex = scanIndex;
     Timestamp = timestamp;
@@ -18,7 +18,7 @@ public class RecordFrame
   public long ScanIndex { get; private set; }
 
   [JsonProperty("timestamp")]
-  public int Timestamp { get; private set; }
+  public float Timestamp { get; private set; }
 
   [JsonProperty("key_code")]
   public VirtualKey KeyCode { get; private set; }
